Moved all pixel types to 'color.hpp'
Considering these types need to be used elsewhere besides textures (such as vertex data) these types have been moved to their own header under 'SH3/types/'
